ImplementationPhase/AOOCT/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
pytest==8.2.0
httpx==0.27.0
numpy==1.26.4
Pillow==10.3.0
matplotlib==3.8.4

